move test classes to package that clearly defines they are needed by the application implementing the library
